requirements.txt
# Bot Pointage - Dépendances
requests>=2.31.0
python-dotenv>=1.0.0
APScheduler>=3.10.0
pyzk>=0.9
